Separate Cron utility logic into its own module.
