+++ coriolis/replica_cron/api.py
@@ -1,5 +1,6 @@
 # Copyright 2017 Cloudbase Solutions Srl
 # All Rights Reserved.
+
 from coriolis.conductor.rpc import client as rpc_client
